+++ test/drawing_test.dart
@@ -3,22 +3,22 @@ import 'dart:io';
 import 'package:flutter_gyw/flutter_gyw.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-main() {
+void main() {
   group("JSON", () {
     test('white screen', () {
-      Drawing drawing = const WhiteScreen();
+      const Drawing drawing = WhiteScreen();
 
       expect(Drawing.fromJson(drawing.toJson()), drawing);
     });
 
     test('Text', () {
-      Drawing drawing = const TextDrawing(text: "Text");
+      const Drawing drawing = TextDrawing(text: "Text");
 
       expect(Drawing.fromJson(drawing.toJson()), drawing);
     });
 
     test('Text with position', () {
-      Drawing drawing = const TextDrawing(
+      const Drawing drawing = TextDrawing(
         text: "Text with position",
         left: 100,
         top: 200,
@@ -28,9 +28,9 @@ main() {
     });
 
     test('Text with font', () {
-      const font = GYWFonts.medium;
+      const GYWFont font = GYWFonts.medium;
 
-      Drawing drawing = const TextDrawing(
+      const Drawing drawing = TextDrawing(
         text: "Text with font",
         font: font,
       );
@@ -41,7 +41,7 @@ main() {
     test('Text with font and position', () {
       const font = GYWFonts.medium;
 
-      Drawing drawing = const TextDrawing(
+      const Drawing drawing = TextDrawing(
         text: "Text with font and position",
         left: 150,
         top: 250,
@@ -52,17 +52,17 @@ main() {
     });
 
     test('Icon', () {
-      const icon = GYWIcons.checkbox;
+      const GYWIcon icon = GYWIcons.checkbox;
 
-      Drawing drawing = const IconDrawing(icon);
+      const Drawing drawing = IconDrawing(icon);
 
       expect(Drawing.fromJson(drawing.toJson()), drawing);
     });
 
     test('Icon with position', () {
-      const icon = GYWIcons.up;
+      const GYWIcon icon = GYWIcons.up;
 
-      Drawing drawing = const IconDrawing(
+      const Drawing drawing = IconDrawing(
         icon,
         left: 120,
         top: 220,
@@ -84,7 +84,8 @@ main() {
   group('Icon assets', () {
     test('All icons', () {
       final assetFolderPath = Platform.environment['UNIT_TEST_ASSETS'];
-      for (GYWIcon icon in GYWIcons.values) {
+
+      for (final GYWIcon icon in GYWIcons.values) {
         expect(
           File("$assetFolderPath/${icon.path}").existsSync(),
           isTrue,